requirements.txt
streamlit
fastapi
uvicorn
torch
torchvision
opencv-python
numpy
Pillow
python-multipart
nest-asyncio
websockets
plotly
pandas
twilio
ultralytics
scikit-image
folium
streamlit-folium
